Add user authentication service and related components
Implemented a new HashService for user password hashing and verification. Introduced AuthenticationDto for user authentication data handling. Also added code to the AuthService to authenticate users and throw validation exceptions when necessary. Created new requests and controllers to handle user authentication, and moved the RegistrationController to the Auth namespace. Updated User entity to store password. Changes also applied to the tests and the 'RegistrationUser.http' file.