Add post creation functionality and update dependencies
This commit adds the ability to create new posts through a modal form in PostScreen, significantly improving the CRUD interactions on post data. It also introduces the Request validation for the said posts. At the same time, it updates the project dependencies such as doctrine/dbal and phpdocumentor/reflection-docblock again to an even more recent version. Lastly, it also includes newly added test cases and environment file to the project.