Remove unused code and providers in AppServiceProvider
This commit simplifies the AppServiceProvider by removing unused imports such as IssueController, AuthMiddleware, and Application. Additionally, unneeded binding of services to their respective contracts were removed. Instead, the services were defined in singleton form which ensures that only a single instance of the service would be created. This refactors the service to have a simpler, cleaner structure without redundant or unused code blocks.